+++ front/lib/api/actions/agent_mcp_action_analytics.ts
