+++ front/pages/api/w/[wId]/labs/mcp_actions/[agentId]/index.ts
@@ -4,12 +4,12 @@ import * as reporter from "io-ts-reporters";
 import { NumberFromString, withFallback } from "io-ts-types";
 import type { NextApiRequest, NextApiResponse } from "next";
 
+import type { MCPAction } from "@app/lib/api/actions/agent_mcp_action_analytics";
+import { AgentMCPActionsAnalytics } from "@app/lib/api/actions/agent_mcp_action_analytics";
 import { getAgentConfiguration } from "@app/lib/api/assistant/configuration";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { getFeatureFlags } from "@app/lib/auth";
-import type { MCPAction } from "@app/lib/resources/agent_mcp_action_resource";
-import { AgentMCPActionsAnalytics } from "@app/lib/resources/agent_mcp_action_resource";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types";
